Store most state client side
This avoids situations where the client can't click "generate"
because the page has been up for too long and the server has dropped
the state. Or any other situation where the server side state is
not available. Instead, we now store the state on the client and
reflect it to the server. There are some exceptions to this, but
they should be recreated in "generate" if necessary.